rename and start refactor on components to fit back end data. app rendering w/o error, signup + login work as expected
