+++ african-marketplace/package.json
@@ -1,4 +1,4 @@
-git checkout combo{
+{
   "name": "african-marketplace",
   "version": "0.1.0",
   "private": true,
