+++ african-marketplace/src/components/SellerItem.jsx
